refactor(gui): extract image saving funcs to logic file
Also, now these functions are using a FrameGrabber
instance as input instead of the FrameNavigation.
