+++ GUI/src/main/kotlin/frameNavigation/FrameNavigation.kt
@@ -3,13 +3,9 @@ package frameNavigation
 import FrameNavigationInterface
 import algorithms.AlignmentElement
 import androidx.compose.runtime.*
-import androidx.compose.ui.graphics.ImageBitmap
-import androidx.compose.ui.graphics.toAwtImage
 import kotlinx.coroutines.*
 import logic.FrameGrabber
 import models.AppState
-import java.awt.Color
-import java.awt.image.BufferedImage
 import kotlin.math.roundToInt
 
 /**
@@ -169,84 +165,6 @@ class FrameNavigation(state: MutableState<AppState>) : FrameNavigationInterface 
         } else {
             diffSequence.take(currentDiffIndex.value).indexOfLast { it != AlignmentElement.PERFECT }
         }
-    }
-
-    /**
-     * Creates a collage of the 3 videos and saves it to a file.
-     * @param outputPath [String] containing the path to save the collage to.
-     * @param border [Int] containing the width of the border between the videos.
-     * @param titleHeight [Int] containing the height of the title.
-     * @param font [java.awt.Font] containing the font to use for the title.
-     * @return [Unit]
-     */
-    fun createCollage(
-        outputPath: String,
-        border: Int = 20,
-        titleHeight: Int = 100,
-        font: java.awt.Font =
-            java.awt.Font(
-                "Arial",
-                java.awt.Font.BOLD,
-                100,
-            ),
-    ) {
-        val width = frameGrabber.width
-        var xOffset = 0
-        // create the collage
-        val collage =
-            BufferedImage(
-                width * 3 + border * 2,
-                frameGrabber.height + titleHeight,
-                BufferedImage.TYPE_INT_RGB,
-            )
-        val g = collage.createGraphics()
-        // fill the background with white
-        g.color = Color.WHITE
-        g.fillRect(0, 0, collage.width, collage.height)
-
-        // draw the images
-        val images =
-            listOf(
-                frameGrabber.getReferenceVideoFrame(currentIndex),
-                frameGrabber.getDiffVideoFrame(currentIndex),
-                frameGrabber.getCurrentVideoFrame(currentIndex),
-            )
-        for (item in images) {
-            val img = item.toAwtImage()
-            g.drawImage(img, xOffset, titleHeight, null)
-            xOffset += width + border
-        }
-
-        // draw the titles
-        g.color = Color.BLACK
-        g.font = font
-        xOffset = 0
-        for (item in listOf("Reference Video", "Diff", "Current Video")) {
-            val metrics = g.fontMetrics
-            val x = (width - metrics.stringWidth(item)) / 2
-            g.drawString(item, x + xOffset, titleHeight - 10)
-            xOffset += width + border
-        }
-
-        // save the collage
-        val file = java.io.File(outputPath)
-        javax.imageio.ImageIO.write(collage, "png", file)
-    }
-
-    /**
-     * Filters all inserted frames.
-     * @return [List]<[ImageBitmap]> containing the inserted frames.
-     */
-    fun getInsertedFrames(): List<ImageBitmap> {
-        val insertedFrames = mutableListOf<ImageBitmap>()
-
-        for (i in diffSequence.indices) {
-            if (diffSequence[i] == AlignmentElement.INSERTION) {
-                insertedFrames.add(frameGrabber.getCurrentVideoFrame(i))
-            }
-        }
-
-        return insertedFrames
     }
 
     fun setOnNavigateCallback(callback: () -> Unit) {
