+++ GUI/src/main/kotlin/ui/components/diffScreen/SaveCollageButton.kt
@@ -9,27 +9,34 @@ import androidx.compose.runtime.MutableState
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.unit.dp
-import frameNavigation.FrameNavigation
+import logic.FrameGrabber
+import logic.createCollage
 import models.AppState
 import ui.components.general.openFileSaverAndGetPath
 
 /**
  * Button to save the collage
  *
- * @param navigator The navigator to use to create the collage
+ * @param frameGrabber The frame grabber to use to get the current frames from
+ * @param diffIndex The index of the current diff
  * @param modifier The modifier for the button
  * @param state The state of the app
  * @return [Unit]
  */
 @Composable
 fun SaveCollageButton(
-    navigator: FrameNavigation,
+    frameGrabber: FrameGrabber,
+    diffIndex: Int,
     modifier: Modifier = Modifier,
     state: MutableState<AppState>,
 ) {
     Button(
         modifier = modifier.padding(8.dp),
-        onClick = { openFileSaverAndGetPath(state.value.saveCollagePath) { path -> saveCollageCallback(navigator, path, state) } },
+        onClick = {
+            openFileSaverAndGetPath(
+                state.value.saveCollagePath,
+            ) { path -> saveCollageCallback(frameGrabber, diffIndex, path, state) }
+        },
     ) {
         Text(
             text = "Save Collage",
@@ -42,12 +49,14 @@ fun SaveCollageButton(
 /**
  * Callback for when the user selects a file to save.
  * Saves the path and creates the collage.
- * @param navigator The navigator to use to create the collage
+ * @param frameGrabber The frame grabber to use to get the current frames from
+ * @param diffIndex The index of the current diff
  * @param path The path to the file to save to
  * @param state The state of the app
  */
 fun saveCollageCallback(
-    navigator: FrameNavigation,
+    frameGrabber: FrameGrabber,
+    diffIndex: Int,
     path: String,
     state: MutableState<AppState>,
 ) {
@@ -57,5 +66,5 @@ fun saveCollageCallback(
         savePath = "$savePath.png"
     }
     state.value.saveCollagePath = savePath
-    navigator.createCollage(savePath)
+    createCollage(frameGrabber = frameGrabber, diffIndex = diffIndex, savePath)
 }
